Add files via upload
Simple linear regression is supervised machine learning algorithm. which is use to find the relationship between one independent variable and one dependent variable. The data-set contain 2 variable one is dependent which is Salary and the other is Experience. so with the help of these data we have to predict the salary based on the Years of Experience.